+++ internal/test/testdata.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -30,7 +31,7 @@ func checkError(t *testing.T, err error) {
 
 var schemeServer *http.Server
 var badServer *http.Server
-var badServerCount int
+var badServerCount atomic.Uint32
 var testStorageDir = "client"
 
 func StartSchemeManagerHttpServer() {
@@ -47,13 +48,12 @@ func StopSchemeManagerHttpServer() {
 }
 
 // StartBadHttpServer starts an HTTP server that times out and returns 500 on the first few times.
-func StartBadHttpServer(count int, timeout time.Duration, success string) {
+func StartBadHttpServer(count uint32, timeout time.Duration, success string) {
 	badServer = &http.Server{Addr: "localhost:48682", Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if badServerCount >= count {
+		if badServerCount.CompareAndSwap(count, 0) {
 			_, _ = fmt.Fprintln(w, success)
-			return
 		} else {
-			badServerCount++
+			badServerCount.Add(1)
 			time.Sleep(timeout)
 		}
 	})}