+++ server/keyshare/keyshareserver/postgresdb.go
@@ -19,10 +19,14 @@ type postgresDB struct {
 	db keyshare.DB
 }
 
-const maxPinTries = 3                  // number of tries allowed on pin before we start with exponential backoff
-const emailTokenValidity = 24          // amount of time user's email validation token is valid (in hours)
-const emailTokenRateLimit = 3          // max number of active tokens per email address within the emailTokenRateLimitDuration
-const emailTokenRateLimitDuration = 60 // amount of time after which tokens become irrelevant for rate limiting (in minutes)
+// Number of tries allowed on pin before we start with exponential backoff
+const maxPinTries = 3
+
+// Max number of active tokens per email address within the emailTokenRateLimitDuration
+const emailTokenRateLimit = 3
+
+// Amount of time after which tokens become irrelevant for rate limiting (in minutes)
+const emailTokenRateLimitDuration = 60
 
 var errTooManyTokens = errors.New("Too many unhandled email tokens for given email address")
 
@@ -44,6 +48,7 @@ func newPostgresDB(connstring string, maxIdleConns, maxOpenConns int, maxIdleTim
 	if err = db.Ping(); err != nil {
 		return nil, errors.Errorf("failed to connect to database: %v", err)
 	}
+
 	return &postgresDB{
 		db: keyshare.DB{DB: db},
 	}, nil
@@ -206,9 +211,9 @@ func (db *postgresDB) addLog(user *User, eventType eventType, param interface{})
 	return err
 }
 
-func (db *postgresDB) addEmailVerification(user *User, emailAddress, token string) error {
-	expiry := time.Now().Add(emailTokenValidity * time.Hour)
-	maxPrevExpiry := expiry.Add(-1 * emailTokenRateLimitDuration * time.Minute)
+func (db *postgresDB) addEmailVerification(user *User, emailAddress, token string, validity int) error {
+	expiry := time.Now().Add(time.Duration(validity) * time.Hour)
+	maxPrevExpiry := expiry.Add(-1 * time.Duration(emailTokenRateLimitDuration) * time.Minute)
 
 	// Check whether rate limiting is necessary
 	amount, err := db.db.ExecCount("SELECT 1 FROM irma.email_verification_tokens WHERE email = $1 AND expiry > $2",